requirements.txt
google-generativeai>=0.8.0
streamlit>=1.40.0
python-dotenv>=1.0.0
python-pptx>=1.0.0
python-docx>=1.1.0
